requirements.txt
asgiref==3.4.1
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==4.0
django-annoying==0.10.6
django-bootstrap==0.2.4
django-heroku==0.3.1
gunicorn==20.1.0
Pillow==8.4.0
psycopg2==2.9.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
